requirements.txt
et-xmlfile==1.0.1
jdcal==1.4.1
lxml==4.6.2
numpy==1.19.5
openpyxl==3.0.6
pandas==1.2.0
Pillow==8.1.0
python-barcode==0.13.1
python-dateutil==2.8.1
python-docx==0.8.10
pytz==2020.5
six==1.15.0
xlrd==2.0.1

pip~=20.3.3
setuptools~=51.3.3
gspread~=3.6.0
oauth2client~=4.1.3